moved these to configuration settings
